@@ -6,13 +6,13 @@ enum VideoType {
 }
 
 class VideoData {
-  final String name;
+  final String? name;
   final String path;
   final VideoType type;
 
   VideoData({
     this.name,
-    this.path,
-    this.type,
+    required this.path,
+    required this.type,
   });
 }
